Migrate integration test to e2e
